devop: move press kit layout under core
