+++ src/core/viewDefault/layouts/ThePressKitLayout.vue
